Do not add new positions to position
If setup as an array it means users want the positions fixed to the
positions they picked, not every possible positions.

For example, if I set position to 'top left' and 'bottom left' it means
I don't want it to display as 'right center' even though
`keepTooltipInside` is set. In that case the code should just pick the
best options among the given positions, not all positions